test: Calculating position from acceleration data
This testing module creates data arrays of acceleration, velocity and position over time (Tx).

While Ax is being filled, we measure Tx as time from beginnig to the i-th Ax mesurement.

Note that in the actual application Vx and Px will be calculated as we read the sensor, and when the arrays get full, the new data will replace the oldest. Using this logic, the arrays could be only 2 variables long, but then we would have a small data log.