feat: first filter
Ignores accelerations with less than 0.0714G and sets velocity to zero if last accelerations were zero.

It was noticed that the acceleration data is always negative because of the slope of the MPU PCB, creating an X axis component of gravity.

Therefore, it is not possible to determine precise X axis acceleration without locking the MPU PCB slope.

Using gyroscope data will be considered to track its slope and correct the gravity influence in X axis acceleration measurements.