feat: gravity filter
The last filter was ineffective because it wasn't noise, but an X-axis component of gravity.

The slight slope of the PCB gave X and Y axis a component of gravity, and the sensor perceived them as acceleration.

This filter:
   - Calculates the X component of gravity
   - Subtracts it from the Ax measurement
   - The result is a sloped acceleration vector, as the sensor's X-axis is tilted. So, it takes multiplies by the cossine of the slope do get real X-axis acceleration

The slope measurement was made using the Kalman filter algorithm, built-in the MPU6050 library.

With this filter, we see that the actual noise in measurement is a few cm/s².