feat: DLPF and SMPLRT_DIV
It was found that this register SMPLRT_DIV (Sample Rate Divider) was limiting the max sample rate indicated in the datasheet (1kHz) to 125Hz (8ms).

It was set to zero, to allow 1kHz reading, and TIM7 was set to 1kHz too.

Also, the DLPF (Digital Low-Pass Filter) was set to 2, which filters noise from measurements. The trade-off is that there is a slight delay (a few miliseconds) between the real life movements and the readings. That should be no problem here.