feat: add Y axis
Now all motion data is correctly measured and processed, in both axis - X and Y.

Px[28] and Py[28] can be used as a real time coordinate of the sensor in a 2D space.

The origin (0, 0) of this space is where the sensor was during calibration. The slope of the sensor has can not changes after calibration. The sensor needs to be at rest during calibration.